mkfs, fsck: change the results for readability
This patch cleans up several printing formats.

+++ include/f2fs_fs.h
@@ -25,7 +25,7 @@ typedef u_int64_t	u64;
 typedef u_int32_t	u32;
 typedef u_int16_t	u16;
 typedef u_int8_t	u8;
-typedef u64		block_t;
+typedef u32		block_t;
 typedef u32		nid_t;
 typedef u8		bool;
 typedef unsigned long	pgoff_t;
